Updated alloc support to work together with external ram properly
