+++ src/prelude.rs
@@ -1,13 +1,12 @@
-pub use embedded_hal::digital::v2::InputPin as _esp32_hal_digital_InputPin;
-pub use embedded_hal::digital::v2::OutputPin as _esp32_hal_digital_OutputPin;
-pub use embedded_hal::digital::v2::StatefulOutputPin as _esp32_hal_digital_StatefulOutputPin;
-pub use embedded_hal::digital::v2::ToggleableOutputPin as _esp32_hal_digital_ToggleableOutputPin;
-pub use embedded_hal::prelude::*;
-
-pub use crate::analog::SensExt as _esp32_hal_analog_SensExt;
-pub use crate::gpio::GpioExt as _esp32_hal_gpio_GpioExt;
+pub use crate::analog::SensExt;
+pub use crate::gpio::GpioExt;
+pub use crate::proc_macros::*;
 pub use crate::units::*;
 
-pub use crate::proc_macros::*;
+pub use embedded_hal::digital::v2::InputPin as _;
+pub use embedded_hal::digital::v2::OutputPin as _;
+pub use embedded_hal::digital::v2::StatefulOutputPin as _;
+pub use embedded_hal::digital::v2::ToggleableOutputPin as _;
+pub use embedded_hal::prelude::*;
 
 pub use xtensa_lx6_rt::entry;
